backend/aloe_api/requirements.txt
django==2.1.9
djangorestframework